# jaboti_backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/select-company": {
      "post": {
        "operationId": "AuthController_selectCompany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/companies": {
      "get": {
        "operationId": "CompaniesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "companies"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "companies"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/companies/{id}": {
      "patch": {
        "operationId": "CompaniesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "companies"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/companies/{id}/join": {
      "post": {
        "operationId": "CompaniesController_join",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "companies"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/departamentos": {
      "get": {
        "operationId": "DepartamentosController_listar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista departamentos da empresa atual",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": 1,
                      "nome": "Atendimento",
                      "createdAt": "2025-01-01T12:00:00Z"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "departamentos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "DepartamentosController_criar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartamentoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "departamentos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/departamentos/{id}": {
      "patch": {
        "operationId": "DepartamentosController_atualizar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartamentoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "departamentos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "DepartamentosController_obter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "departamentos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "DepartamentosController_remover",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "departamentos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/departamentos/{id}/membros": {
      "get": {
        "operationId": "DepartamentosController_listarMembros",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista membros (pessoas) do departamento",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": 10,
                      "name": "João Operador",
                      "type": "USUARIO"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "departamentos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "DepartamentosController_adicionarMembros",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMembrosDepartamentoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Adiciona membros ao departamento",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "mensagem": "2 atendente(s) adicionado(s)",
                    "adicionados": [
                      1,
                      2
                    ],
                    "jaAdicionados": [
                      3
                    ],
                    "naoEncontrados": [
                      99
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Somente atendentes (USUARIO) podem ser adicionados ao departamento"
          }
        },
        "tags": [
          "departamentos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/departamentos/{id}/membros/{pessoaId}": {
      "delete": {
        "operationId": "DepartamentosController_removerMembro",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pessoaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remove membro do departamento",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "removido": true
                  }
                }
              }
            }
          }
        },
        "tags": [
          "departamentos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Jaboti API",
    "description": "API multiempresa de atendimento Jaboti",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "RefreshDto": {
        "type": "object",
        "properties": {}
      },
      "SelectCompanyDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCompanyDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {}
      },
      "CreateDepartamentoDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Atendimento",
            "description": "Nome do departamento (exibido nas telas)"
          }
        },
        "required": [
          "nome"
        ]
      },
      "UpdateDepartamentoDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Suporte Nível 2",
            "description": "Novo nome do departamento"
          }
        }
      },
      "AddMembrosDepartamentoDto": {
        "type": "object",
        "properties": {
          "pessoaIds": {
            "example": [
              3,
              5,
              9
            ],
            "description": "IDs de pessoas (USUARIOS) a adicionar ao departamento",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "pessoaIds"
        ]
      },
...
    }
  }
}